LayoutTests/platform/win-future/editing/spelling/spelling-backspace-between-lines-expected.txt

For Bug 41423: Spelling marker should remain after hitting a backspace key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS firstLineText() is 'OK'
PASS sel.anchorNode.data is 'OK zz OK'
PASS internals.hasSpellingMarker(document, 3, 2) is true
PASS sel.anchorNode.data is 'OKOK zz OK'
PASS firstLineText() is 'OKOK zz OK'
FAIL internals.hasSpellingMarker(document, 5, 2) should be true. Was false.
PASS firstLineText() is 'OK'
PASS sel.anchorNode.data is 'OK zz OK'
PASS internals.hasSpellingMarker(document, 3, 2) is true
PASS firstLineText() is 'OKOK zz OK'
PASS sel.anchorNode.data is 'OKOK zz OK'
FAIL internals.hasSpellingMarker(document, 5, 2) should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
